Add new component for collaboration banner
