+++ src/styles/collaboration-banner.scss
@@ -1,0 +1,9 @@
+.collaboration-banner {
+    .columns {
+        margin-top: 1.5rem;
+
+        .button {
+          width: 100%;
+        }
+    }
+}
