+++ src/styles/partners.scss
@@ -3,14 +3,6 @@
     .hero-body {
       width: 100%;
       max-width: 720px;
-
-      .columns {
-        margin-top: 1.5rem;
-
-        .button {
-          width: 100%;
-        }
-      }
     }
   }
   .logos {
